pyproject.toml
[project]
name = "copilot-telegram-bot"
version = "0.1.0"
description = "Using GitHub Copilot on your Telegram"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.0",
    "python-dotenv>=1.0.0",
    "github-copilot-sdk==0.2.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]